Add CAB for deleting ships, fixes #50
Currently the workflow is somewhat awkward, involving a long-press to
trigger a contextual action bar with delete.

This CAB could also be fitted with an option for other rare, ship
level actions, such as swapping out the Ship itself.

Also, moved onepane->twopane alias to be behind sw qualifier. This way,
we don't try to convert a single pane back stack to and from a multi
pane back stack.