Fix tests on Android
Fixed cost calculation for flagships and fleet captains so that the resource cost is included even if the item isn't yet installed into the squad.

Fixed the load test for the new items. The specials test doesn't work yet as the officers are missing and the upgrades and the data items need to be sorted the same way for the matching to work. See the Objective-C implementation for an example.

Added officers to the list of items to be loaded so there would be less logging.